maha(ClassPrediction)	R Documentation

Compute the (squared) Mahalanobis distance between two groups of vectors

Description

The Mahalanobis distance between two groups of vectors

Usage


maha(data, groups, method = "mve")

Arguments

data 	 A matrix with columns representing features (or variables) and rows representing independent samples

groups 	 A factor or logical vector with length equal to the number of rows (samples) in the data matrix

method 	 A character string determing the method that should be used to estimate thecovariance matrix. The default value of "mve" uses the cov.mve function from the MASS package. The other valid option is "var", which uses the var function from the standard stats package.

Details

Stuff should go here

Value

Returns a numeric vector of length 1.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

cov.mve, var

Examples


nFeatures <- 40
nSamples <- 2*10
dataset <- matrix(rnorm(nSamples*nFeatures), ncol=nSamples)
groups <- factor(rep(c("A", "B"), each=10))
maha(dataset, groups)

[Package ClassPrediction version 2.9.0 Index]